refactor: encryption 수행 패키지 utils 패키지로, static으로 재구성
