+++ src/main/java/com/lgmpjt/websnsspringboot/utils/SHA256.java
@@ -1,4 +1,4 @@
-package com.lgmpjt.websnsspringboot.encryption.service;
+package com.lgmpjt.websnsspringboot.utils;
 
 import org.springframework.stereotype.Service;
 
@@ -10,14 +10,14 @@ public class SHA256 {
 
 	private final static String SHA_256 = "SHA-256";
 
-	public String encrypt(String text) throws NoSuchAlgorithmException {
+	public static String encrypt(String text) throws NoSuchAlgorithmException {
 		MessageDigest messageDigest = MessageDigest.getInstance(SHA_256);
 		messageDigest.update(text.getBytes());
 
 		return bytesToHex(messageDigest.digest());
 	}
 
-	private String bytesToHex(byte[] bytes) {
+	private static String bytesToHex(byte[] bytes) {
 		StringBuilder sb = new StringBuilder();
 		for (byte b : bytes) {
 			sb.append(String.format("%02x", b));
